+++ cmd2/cmd2.py
@@ -700,10 +700,7 @@ class Cmd(cmd.Cmd):
         elif callable(parser_builder):
             parser = parser_builder()
         elif isinstance(parser_builder, argparse.ArgumentParser):
-            if sys.version_info >= (3, 6, 4):
-                parser = copy.deepcopy(parser_builder)
-            else:  # pragma: no cover
-                parser = parser_builder
+            parser = copy.deepcopy(parser_builder)
         return parser
 
     def _register_command_parser(self, command: str, command_method: Callable[..., Any]) -> None:
@@ -778,7 +775,7 @@ class Cmd(cmd.Cmd):
             cmdset.on_unregister()
             self._unregister_subcommands(cmdset)
 
-            methods = inspect.getmembers(
+            methods: List[Tuple[str, Callable[[Any], Any]]] = inspect.getmembers(
                 cmdset,
                 predicate=lambda meth: isinstance(meth, Callable)  # type: ignore[arg-type]
                 and hasattr(meth, '__name__')
@@ -809,7 +806,7 @@ class Cmd(cmd.Cmd):
             self._installed_command_sets.remove(cmdset)
 
     def _check_uninstallable(self, cmdset: CommandSet) -> None:
-        methods = inspect.getmembers(
+        methods: List[Tuple[str, Callable[[Any], Any]]] = inspect.getmembers(
             cmdset,
             predicate=lambda meth: isinstance(meth, Callable)  # type: ignore[arg-type]
             and hasattr(meth, '__name__')
@@ -1223,7 +1220,6 @@ class Cmd(cmd.Cmd):
         msg: Any = '',
         *,
         end: str = '\n',
-        apply_style: bool = True,
         paged: bool = False,
         chop: bool = False,
     ) -> None:
@@ -1231,16 +1227,10 @@ class Cmd(cmd.Cmd):
 
         :param msg: object to print
         :param end: string appended after the end of the message, default a newline
-        :param apply_style:
-            If True, then ansi.style_warning will be applied to the message text. Set to False in cases
-            where the message text already has the desired style. Defaults to True.
-
-            .. deprecated: 2.4.4
-                Use :meth:`~cmd2.Cmd.print_to` instead to print to stderr without style applied.
         :param paged: If True, pass the output through the configured pager.
         :param chop: If paged is True, True to truncate long lines or False to wrap long lines.
         """
-        self.print_to(sys.stderr, msg, end=end, style=ansi.style_warning if apply_style else None, paged=paged, chop=chop)
+        self.print_to(sys.stderr, msg, end=end, style=ansi.style_warning, paged=paged, chop=chop)
 
     def pfailure(
         self,
@@ -2415,13 +2405,13 @@ class Cmd(cmd.Cmd):
         return [topic for topic in all_topics if topic not in self.hidden_commands and topic not in self.disabled_commands]
 
     # noinspection PyUnusedLocal
-    def sigint_handler(self, signum: int, _: FrameType) -> None:
+    def sigint_handler(self, signum: int, _: Optional[FrameType]) -> None:
         """Signal handler for SIGINTs which typically come from Ctrl-C events.
 
-        If you need custom SIGINT behavior, then override this function.
+        If you need custom SIGINT behavior, then override this method.
 
         :param signum: signal number
-        :param _: required param for signal handlers
+        :param _: the current stack frame or None
         """
         if self._cur_pipe_proc_reader is not None:
             # Pass the SIGINT to the current pipe process
@@ -2436,6 +2426,23 @@ class Cmd(cmd.Cmd):
                     raise_interrupt = not command_set.sigint_handler()
             if raise_interrupt:
                 self._raise_keyboard_interrupt()
+
+    def termination_signal_handler(self, signum: int, _: Optional[FrameType]) -> None:
+        """
+        Signal handler for SIGHUP and SIGTERM. Only runs on Linux and Mac.
+
+        SIGHUP - received when terminal window is closed
+        SIGTERM - received when this app has been requested to terminate
+
+        The basic purpose of this method is to call sys.exit() so our exit handler will run
+        and save the persistent history file. If you need more complex behavior like killing
+        threads and performing cleanup, then override this method.
+
+        :param signum: signal number
+        :param _: the current stack frame or None
+        """
+        # POSIX systems add 128 to signal numbers for the exit code
+        sys.exit(128 + signum)
 
     def _raise_keyboard_interrupt(self) -> None:
         """Helper function to raise a KeyboardInterrupt"""
@@ -3335,19 +3342,14 @@ class Cmd(cmd.Cmd):
     #############################################################
 
     # Top-level parser for alias
-    @staticmethod
-    def _build_alias_parser() -> argparse.ArgumentParser:
-        alias_description = (
-            "Manage aliases\n" "\n" "An alias is a command that enables replacement of a word by another string."
-        )
-        alias_epilog = "See also:\n" "  macro"
-        alias_parser = argparse_custom.DEFAULT_ARGUMENT_PARSER(description=alias_description, epilog=alias_epilog)
-        alias_subparsers = alias_parser.add_subparsers(dest='subcommand', metavar='SUBCOMMAND')
-        alias_subparsers.required = True
-        return alias_parser
+    alias_description = "Manage aliases\n" "\n" "An alias is a command that enables replacement of a word by another string."
+    alias_epilog = "See also:\n" "  macro"
+    alias_parser = argparse_custom.DEFAULT_ARGUMENT_PARSER(description=alias_description, epilog=alias_epilog)
+    alias_subparsers = alias_parser.add_subparsers(dest='subcommand', metavar='SUBCOMMAND')
+    alias_subparsers.required = True
 
     # Preserve quotes since we are passing strings to other commands
-    @with_argparser(_build_alias_parser, preserve_quotes=True)
+    @with_argparser(alias_parser, preserve_quotes=True)
     def do_alias(self, args: argparse.Namespace) -> None:
         """Manage aliases"""
         # Call handler for whatever subcommand was selected
@@ -5457,11 +5459,18 @@ class Cmd(cmd.Cmd):
         if not threading.current_thread() is threading.main_thread():
             raise RuntimeError("cmdloop must be run in the main thread")
 
-        # Register a SIGINT signal handler for Ctrl+C
+        # Register signal handlers
         import signal
 
         original_sigint_handler = signal.getsignal(signal.SIGINT)
-        signal.signal(signal.SIGINT, self.sigint_handler)  # type: ignore
+        signal.signal(signal.SIGINT, self.sigint_handler)
+
+        if not sys.platform.startswith('win'):
+            original_sighup_handler = signal.getsignal(signal.SIGHUP)
+            signal.signal(signal.SIGHUP, self.termination_signal_handler)
+
+            original_sigterm_handler = signal.getsignal(signal.SIGTERM)
+            signal.signal(signal.SIGTERM, self.termination_signal_handler)
 
         # Grab terminal lock before the command line prompt has been drawn by readline
         self.terminal_lock.acquire()
@@ -5495,8 +5504,12 @@ class Cmd(cmd.Cmd):
         # This will also zero the lock count in case cmdloop() is called again
         self.terminal_lock.release()
 
-        # Restore the original signal handler
+        # Restore original signal handlers
         signal.signal(signal.SIGINT, original_sigint_handler)
+
+        if not sys.platform.startswith('win'):
+            signal.signal(signal.SIGHUP, original_sighup_handler)
+            signal.signal(signal.SIGTERM, original_sigterm_handler)
 
         return self.exit_code