+++ cmd2/decorators.py
@@ -209,6 +209,7 @@ def _set_parser_prog(parser: argparse.ArgumentParser, prog: str) -> None:
     """
     # Set the prog value for this parser
     parser.prog = prog
+    req_args: List[str] = []
 
     # Set the prog value for the parser's subcommands
     for action in parser._actions:
@@ -233,12 +234,19 @@ def _set_parser_prog(parser: argparse.ArgumentParser, prog: str) -> None:
                 if subcmd_parser in processed_parsers:
                     continue
 
-                subcmd_prog = parser.prog + ' ' + subcmd_name
+                subcmd_prog = parser.prog
+                if req_args:
+                    subcmd_prog += " " + " ".join(req_args)
+                subcmd_prog += " " + subcmd_name
                 _set_parser_prog(subcmd_parser, subcmd_prog)
                 processed_parsers.append(subcmd_parser)
 
             # We can break since argparse only allows 1 group of subcommands per level
             break
+
+        # Need to save required args so they can be prepended to the subcommand usage
+        elif action.required:
+            req_args.append(action.dest)
 
 
 #: Function signature for a Command Function that uses an argparse.ArgumentParser to process user input
@@ -266,8 +274,7 @@ def with_argparser(
     ns_provider: Optional[Callable[..., argparse.Namespace]] = None,
     preserve_quotes: bool = False,
     with_unknown_args: bool = False,
-) -> Callable[[ArgparseCommandFunc[CommandParent]], RawCommandFuncOptionalBoolReturn[CommandParent]]:
-    ...  # pragma: no cover
+) -> Callable[[ArgparseCommandFunc[CommandParent]], RawCommandFuncOptionalBoolReturn[CommandParent]]: ...  # pragma: no cover
 
 
 @overload
@@ -277,8 +284,7 @@ def with_argparser(
     ns_provider: Optional[Callable[..., argparse.Namespace]] = None,
     preserve_quotes: bool = False,
     with_unknown_args: bool = False,
-) -> Callable[[ArgparseCommandFunc[CommandParent]], RawCommandFuncOptionalBoolReturn[CommandParent]]:
-    ...  # pragma: no cover
+) -> Callable[[ArgparseCommandFunc[CommandParent]], RawCommandFuncOptionalBoolReturn[CommandParent]]: ...  # pragma: no cover
 
 
 def with_argparser(
@@ -418,8 +424,7 @@ def as_subcommand_to(
     *,
     help: Optional[str] = None,
     aliases: Optional[List[str]] = None,
-) -> Callable[[ArgparseCommandFunc[CommandParent]], ArgparseCommandFunc[CommandParent]]:
-    ...  # pragma: no cover
+) -> Callable[[ArgparseCommandFunc[CommandParent]], ArgparseCommandFunc[CommandParent]]: ...  # pragma: no cover
 
 
 @overload
@@ -430,8 +435,7 @@ def as_subcommand_to(
     *,
     help: Optional[str] = None,
     aliases: Optional[List[str]] = None,
-) -> Callable[[ArgparseCommandFunc[CommandParent]], ArgparseCommandFunc[CommandParent]]:
-    ...  # pragma: no cover
+) -> Callable[[ArgparseCommandFunc[CommandParent]], ArgparseCommandFunc[CommandParent]]: ...  # pragma: no cover
 
 
 def as_subcommand_to(